- Remove multiple instances of HttpClient in mock tests
